nlenc: initial commit, move integer encoding/decoding functions
