+++ message.go
@@ -2,6 +2,8 @@ package netlink
 
 import (
 	"errors"
+
+	"github.com/mdlayher/netlink/nlenc"
 )
 
 // Various errors which may occur when attempting to marshal or unmarshal
@@ -122,11 +124,11 @@ func (m Message) MarshalBinary() ([]byte, error) {
 
 	b := make([]byte, ml)
 
-	PutUint32(b[0:4], m.Header.Length)
-	PutUint16(b[4:6], uint16(m.Header.Type))
-	PutUint16(b[6:8], uint16(m.Header.Flags))
-	PutUint32(b[8:12], m.Header.Sequence)
-	PutUint32(b[12:16], m.Header.PID)
+	nlenc.PutUint32(b[0:4], m.Header.Length)
+	nlenc.PutUint16(b[4:6], uint16(m.Header.Type))
+	nlenc.PutUint16(b[6:8], uint16(m.Header.Flags))
+	nlenc.PutUint32(b[8:12], m.Header.Sequence)
+	nlenc.PutUint32(b[12:16], m.Header.PID)
 	copy(b[16:], m.Data)
 
 	return b, nil
@@ -142,15 +144,15 @@ func (m *Message) UnmarshalBinary(b []byte) error {
 	}
 
 	// Don't allow misleading length
-	m.Header.Length = Uint32(b[0:4])
+	m.Header.Length = nlenc.Uint32(b[0:4])
 	if int(m.Header.Length) != len(b) {
 		return errShortMessage
 	}
 
-	m.Header.Type = HeaderType(Uint16(b[4:6]))
-	m.Header.Flags = HeaderFlags(Uint16(b[6:8]))
-	m.Header.Sequence = Uint32(b[8:12])
-	m.Header.PID = Uint32(b[12:16])
+	m.Header.Type = HeaderType(nlenc.Uint16(b[4:6]))
+	m.Header.Flags = HeaderFlags(nlenc.Uint16(b[6:8]))
+	m.Header.Sequence = nlenc.Uint32(b[8:12])
+	m.Header.PID = nlenc.Uint32(b[12:16])
 	m.Data = b[16:]
 
 	return nil
@@ -169,7 +171,7 @@ func checkMessage(m Message) error {
 		return errShortErrorMessage
 	}
 
-	if c := getInt32(m.Data[0:4]); c != success {
+	if c := nlenc.Int32(m.Data[0:4]); c != success {
 		// Error code is a negative integer, convert it into
 		// an OS-specific system call error
 		return newError(-1 * int(c))
