+++ nlenc/doc.go
@@ -1,0 +1,3 @@
+// Package nlenc implements encoding and decoding functions for netlink
+// messages and attributes.
+package nlenc
